:hammer: Rename FlowTypes files to types
