+++ ern-cauldron-api/test/CauldronApi-test.ts
@@ -3,7 +3,7 @@ import sinon from 'sinon'
 import fs from 'fs'
 import { NativeApplicationDescriptor, PackagePath } from 'ern-core'
 import { doesThrow, fixtures } from 'ern-util-dev'
-import { CauldronCodePushEntry } from '../src/FlowTypes'
+import { CauldronCodePushEntry } from '../src/types'
 import CauldronApi from '../src/CauldronApi'
 import EphemeralFileStore from '../src/EphemeralFileStore'
 import InMemoryDocumentStore from '../src/InMemoryDocumentStore'
