+++ ern-cauldron-api/test/InMemoryDocumentStore-test.ts
@@ -3,7 +3,7 @@ import { fixtures } from 'ern-util-dev'
 import fs from 'fs'
 import path from 'path'
 import InMemoryDocumentStore from '../src/InMemoryDocumentStore'
-import { Cauldron } from '../src/FlowTypes'
+import { Cauldron } from '../src/types'
 import sinon from 'sinon'
 const sandbox = sinon.createSandbox()
